em nộp bài đã fix ạ

+++ src/ss12_java_collection_framwork/practice_java_collection_framework/model/Product.java
@@ -3,12 +3,12 @@ package ss12_java_collection_framwork.practice_java_collection_framework.model;
 public class Product {
     private int id;
     private String name;
-    private int price;
+    private double price;
 
     public Product() {
     }
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,11 +30,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
